api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-pubsub==2.18.4
google-cloud-firestore==2.13.1
pydantic==2.5.2
